% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{reference_portfolios}
\alias{reference_portfolios}
\title{Synthetic integration portfolios}
\usage{
reference_portfolios()
}
\value{
A list containing named dependency nodes, explicit contexts and
independently specified expected values for the reference portfolios.
}
\description{
Synthetic integration portfolios
}
\examples{
length(reference_portfolios()$portfolios)
}
\seealso{
\code{\link[=reference_cases]{reference_cases()}}, \code{\link[=rule_profiles]{rule_profiles()}}
}
